fix: rename type files as lowercase (fixes #4028)
