+++ packages/vite/src/node/index.ts
@@ -82,4 +82,4 @@ export type { FSWatcher, WatchOptions } from 'types/chokidar'
 export type { Terser } from 'types/terser'
 export type { CleanCSS } from 'types/clean-css'
 export type { RollupCommonJSOptions } from 'types/commonjs'
-export type { RollupDynamicImportVarsOptions } from 'types/dynamicImportVars'
+export type { RollupDynamicImportVarsOptions } from 'types/dynamic-import-vars'
