+++ packages/vite/types/custom-event.d.ts
