+++ packages/vite/types/dynamic-import-vars.d.ts
